requirements.txt
beautifulsoup4
requests
lxml
pydantic==2.7.1
pandas==2.2.2
polars==0.20.22
psycopg2==2.9.9
python-dotenv==1.0.1